feat: add navbar with username
