+++ src/components/NavBar/NavBar.d.ts
@@ -1,0 +1,3 @@
+export type NavBarProps = {
+  logout: () => void
+}
