+++ src/components/NavBar/index.tsx
@@ -1,0 +1,116 @@
+import * as React from "react";
+import {useEffect, useState} from "react";
+import AppBar from "@mui/material/AppBar";
+import Box from "@mui/material/Box";
+import Toolbar from "@mui/material/Toolbar";
+import IconButton from "@mui/material/IconButton";
+import Typography from "@mui/material/Typography";
+import Menu from "@mui/material/Menu";
+import Container from "@mui/material/Container";
+import Avatar from "@mui/material/Avatar";
+import Tooltip from "@mui/material/Tooltip";
+import MenuItem from "@mui/material/MenuItem";
+import TaskAlt from "@mui/icons-material/TaskAlt";
+import Logout from "@mui/icons-material/Logout";
+import axios from "axios";
+
+import {NavBarProps} from './NavBar';
+
+const NavBar = (props:NavBarProps) => {
+
+  const { logout } = props;
+
+  const [userData, setUserData] = useState<null|{
+    nome:string;
+    login:string;
+    admin:boolean;
+  }>(null);
+
+  useEffect(() => {
+    axios.get('http://localhost:3000/usuarios/autenticado').then((response) => {
+      console.log('xxx', response.data)
+      setUserData(response.data.usuario)
+    })
+  }, [])
+
+  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
+    null
+  );
+
+  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
+    setAnchorElUser(event.currentTarget);
+  };
+
+  const handleCloseUserMenu = () => {
+    setAnchorElUser(null);
+  };
+
+
+
+  return (
+    <AppBar position="static">
+      <Container maxWidth="xl">
+        <Toolbar disableGutters sx={{
+          justifyContent: 'space-between'
+        }}>
+          <Box sx={{
+            display: 'flex',
+            alignItems: 'center'
+          }}>
+            <TaskAlt sx={{ display: "flex", mr: 2 }} />
+
+            <Typography
+              variant="h6"
+              noWrap
+              component="a"
+              href="/"
+              sx={{
+                mr: 2,
+                display: "flex",
+                fontFamily: "monospace",
+                fontWeight: 700,
+                letterSpacing: ".3rem",
+                color: "inherit",
+                textDecoration: "none",
+              }}
+            >
+              TaFeito
+            </Typography>
+            {userData ? <Typography> Bem vindo: {userData.nome}  </Typography> : null}
+          </Box>
+
+          <Box sx={{ flexGrow: 0 }}>
+            <Tooltip title="Open settings">
+              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
+                <Avatar alt="Usuario Anonimo" />
+              </IconButton>
+            </Tooltip>
+            <Menu
+              sx={{ mt: "45px" }}
+              id="menu-appbar"
+              anchorEl={anchorElUser}
+              anchorOrigin={{
+                vertical: "top",
+                horizontal: "right",
+              }}
+              keepMounted
+              transformOrigin={{
+                vertical: "top",
+                horizontal: "right",
+              }}
+              open={Boolean(anchorElUser)}
+              onClose={handleCloseUserMenu}
+            >
+              <MenuItem onClick={logout}>
+                <Logout sx={{ display: "flex", mr: 2 }} />
+                <Typography textAlign="center">Sair</Typography>
+              </MenuItem>
+            </Menu>
+          </Box>
+        </Toolbar>
+      </Container>
+    </AppBar>
+  );
+};
+
+export default NavBar;
